+++ src/PhpWord/Writer/Word2007/Footer.php
@@ -23,12 +23,9 @@ class Footer extends Base
      */
     public function writeFooter(FooterElement $footer)
     {
-        // Create XML writer
         $xmlWriter = $this->getXmlWriter();
 
-        // XML header
         $xmlWriter->startDocument('1.0', 'UTF-8', 'yes');
-
         $xmlWriter->startElement('w:ftr');
         $xmlWriter->writeAttribute('xmlns:ve', 'http://schemas.openxmlformats.org/markup-compatibility/2006');
         $xmlWriter->writeAttribute('xmlns:o', 'urn:schemas-microsoft-com:office:office');
@@ -42,9 +39,8 @@ class Footer extends Base
 
         $this->writeContainerElements($xmlWriter, $footer);
 
-        $xmlWriter->endElement();
+        $xmlWriter->endElement(); // w:ftr
 
-        // Return
         return $xmlWriter->getData();
     }
 }